v1.1.5

+++ clean.bat
@@ -1,7 +1,10 @@
 @echo off
 echo Cleaning project...
-rmdir /s /q build
-rmdir /s /q __pycache__
-del /q *.spec
-rmdir /s /q venv
+
+if exist build rmdir /s /q build
+if exist __pycache__ rmdir /s /q __pycache__
+if exist *.spec del /q *.spec
+if exist venv rmdir /s /q venv
+
 echo Clean complete.
+exit /b 0
